refactor: separate factories in folders
