+++ src/main/factories/controllers/signup/makeSignupController.ts
@@ -1,0 +1,11 @@
+import { SignUpController } from '../../../../presentation/controllers/signup/SignUpController'
+import { Controller } from '../../../../presentation/protocols'
+import { makeLogControllerDecorator } from '../../decorators/makeLogControllerDecorator'
+import { makeDBAddAccount } from '../../useCases/addAccount/makeDBAddAccount'
+import { makeDBAuthentication } from '../../useCases/authentication/makeDBAuthentication'
+import { makeSignupValidation } from './makeSignupValidation'
+
+export function makeSignupController (): Controller {
+  const signupController = new SignUpController(makeDBAddAccount(), makeSignupValidation(), makeDBAuthentication())
+  return makeLogControllerDecorator(signupController)
+}
