+++ src/main/factories/controllers/signup/makeSignupValidation.spec.ts
@@ -1,14 +1,10 @@
-import {
-  RequiredFieldValidation,
-  CompareFieldsValidation,
-  EmailValidation,
-  ValidationComposite
-} from '../../../presentation/helpers/validators'
-import { EmailValidator } from '../../../presentation/protocols/EmailValidator'
-import { Validation } from '../../../presentation/protocols/Validation'
+
+import { RequiredFieldValidation, CompareFieldsValidation, EmailValidation, ValidationComposite } from '../../../../presentation/helpers/validators'
+import { EmailValidator } from '../../../../presentation/protocols/EmailValidator'
+import { Validation } from '../../../../presentation/protocols/Validation'
 import { makeSignupValidation } from './makeSignupValidation'
 
-jest.mock('../../../presentation/helpers/validators/ValidationComposite')
+jest.mock('../../../../presentation/helpers/validators/ValidationComposite')
 
 function makeEmailValidator (): EmailValidator {
   class EmailValidatorStub implements EmailValidator {
