+++ src/main/factories/decorators/makeLogControllerDecorator.ts
@@ -1,0 +1,8 @@
+import { LogMongoRepository } from '../../../infra/db/mongodb/log/LogMongoRepository'
+import { Controller } from '../../../presentation/protocols'
+import { LogControllerDecorator } from '../../decorators/LogControllerDecorator'
+
+export function makeLogControllerDecorator (controller:Controller):Controller {
+  const logErrorRepository = new LogMongoRepository()
+  return new LogControllerDecorator(controller, logErrorRepository)
+}
